+++ CHANGELOG.md
@@ -189,6 +189,9 @@
 - Add logging output of parse CLI function
 - Hotfix Poutyne version handling
 
-## 0.7.3
+## dev
 - Add freeze layers parameters to freeze layers during retraining
 - Improve parsed address print
+- Bug-fix #124: comma-separated list without whitespace in CSVDatasetContainer
+- Add a report when addresses to parse and tags list len differ
+- Add an example on how to fine-tune using our CSVDatasetContainer